完成作业3修改

+++ src/tensor/tensor.cpp
@@ -233,7 +233,7 @@ tensor_t Tensor::view(const std::vector<size_t> &shape) const {
     // 计算新的步长
     std::vector<ptrdiff_t> new_strides(shape.size());
     ptrdiff_t stride = 1;
-    for (size_t i = shape.size() - 1; i >= 0; --i) {
+    for (int i = shape.size() - 1; i >= 0; --i) {
         new_strides[i] = stride;
         stride *= shape[i];
     }
